Add files via upload
This code is intended as developmental support for new researchers to the BCI field. Please cite the following publication if this code is used for research purposes:

Harnarinesingh, R.E.S. and C.S. Syan (2018), Investigation of the Mirrored-Word Reading Paradigm for BCI Implementation, Biomedical Engineering / Biomedizinische Technik, vol. 36, Ahead of Print